Refine moss_soundeffect_v2 layout and docs
