+++ clis/moss_sound_effect_app.py
@@ -1,161 +1,85 @@
+"""Gradio demo for the moss_soundeffect_v2 subproject."""
+
 import argparse
 import functools
-import importlib.util
+import os
+import sys
 import time
+from pathlib import Path
 
 import gradio as gr
 import numpy as np
 import torch
-from transformers import AutoModel, AutoProcessor
 
-# Disable the broken cuDNN SDPA backend
-torch.backends.cuda.enable_cudnn_sdp(False)
-# Keep these enabled as fallbacks
-torch.backends.cuda.enable_flash_sdp(True)
-torch.backends.cuda.enable_mem_efficient_sdp(True)
-torch.backends.cuda.enable_math_sdp(True)
+_HERE = Path(__file__).resolve().parent
+_REPO_ROOT = _HERE.parent
+sys.path.insert(0, str(_REPO_ROOT))
 
-MODEL_PATH = "OpenMOSS-Team/MOSS-SoundEffect"
-DEFAULT_ATTN_IMPLEMENTATION = "auto"
-DEFAULT_MAX_NEW_TOKENS = 4096
-TOKENS_PER_SECOND = 12.5
+from moss_soundeffect_v2 import MossSoundEffectPipeline
+
+
+DEFAULT_MODEL_DIR = os.environ.get("SOUNDEFFECT_MODEL_DIR", "/path/to/SoundEffect-v2-hf")
+DEFAULT_DEVICE = os.environ.get("SOUNDEFFECT_DEVICE", "cuda")
+MAX_INFERENCE_SECONDS = 30
 
 
 @functools.lru_cache(maxsize=1)
-def load_backend(model_path: str, device_str: str, attn_implementation: str):
+def load_backend(model_dir: str, device_str: str):
     device = torch.device(device_str if torch.cuda.is_available() else "cpu")
-    dtype = torch.bfloat16 if device.type == "cuda" else torch.float32
-    resolved_attn_implementation = resolve_attn_implementation(
-        requested=attn_implementation,
-        device=device,
-        dtype=dtype,
+    pipe = MossSoundEffectPipeline.from_pretrained(
+        model_dir,
+        torch_dtype=torch.bfloat16,
+        device=str(device),
     )
-
-    processor = AutoProcessor.from_pretrained(
-        model_path,
-        trust_remote_code=True,
-    )
-    if hasattr(processor, "audio_tokenizer"):
-        processor.audio_tokenizer = processor.audio_tokenizer.to(device)
-
-    model_kwargs = {
-        "trust_remote_code": True,
-        "torch_dtype": dtype,
-    }
-    if resolved_attn_implementation:
-        model_kwargs["attn_implementation"] = resolved_attn_implementation
-
-    model = AutoModel.from_pretrained(model_path, **model_kwargs).to(device)
-    model.eval()
-
-    sample_rate = int(getattr(processor.model_config, "sampling_rate", 24000))
-    return model, processor, device, sample_rate
-
-
-def resolve_attn_implementation(requested: str, device: torch.device, dtype: torch.dtype) -> str | None:
-    requested_norm = (requested or "").strip().lower()
-
-    if requested_norm in {"none"}:
-        return None
-
-    if requested_norm not in {"", "auto"}:
-        return requested
-
-    # Prefer FlashAttention 2 when package + device conditions are met.
-    if (
-        device.type == "cuda"
-        and importlib.util.find_spec("flash_attn") is not None
-        and dtype in {torch.float16, torch.bfloat16}
-    ):
-        major, _ = torch.cuda.get_device_capability(device)
-        if major >= 8:
-            return "flash_attention_2"
-
-    # CUDA fallback: use PyTorch SDPA kernels.
-    if device.type == "cuda":
-        return "sdpa"
-
-    # CPU fallback.
-    return "eager"
-
-
-def build_conversation(ambient_sound: str, duration_seconds: float, processor):
-    ambient_sound = (ambient_sound or "").strip()
-    if not ambient_sound:
-        raise ValueError("Please enter an ambient sound description.")
-
-    expected_tokens = max(1, int(float(duration_seconds) * TOKENS_PER_SECOND))
-    user_kwargs = {
-        "ambient_sound": ambient_sound,
-        "tokens": expected_tokens,
-    }
-
-    return [[processor.build_user_message(**user_kwargs)]], expected_tokens
+    return pipe, device
 
 
 def run_inference(
-    ambient_sound: str,
-    duration_seconds: float,
-    temperature: float,
-    top_p: float,
-    top_k: int,
-    repetition_penalty: float,
-    max_new_tokens: int,
-    model_path: str,
+    prompt: str,
+    seconds: float,
+    steps: int,
+    cfg_scale: float,
+    sigma_shift: float,
+    seed: int,
+    model_dir: str,
     device: str,
-    attn_implementation: str,
 ):
+    if not (prompt or "").strip():
+        raise ValueError("Please enter a prompt describing the audio you want to generate.")
+
+    seconds = round(float(seconds), 1)
+    if seconds <= 0:
+        raise ValueError("Duration must be greater than 0.")
+    if seconds > MAX_INFERENCE_SECONDS:
+        raise ValueError(f"Duration must be no greater than {MAX_INFERENCE_SECONDS}s.")
+
     started_at = time.monotonic()
-    model, processor, torch_device, sample_rate = load_backend(
-        model_path=model_path,
-        device_str=device,
-        attn_implementation=attn_implementation,
+    pipe, torch_device = load_backend(model_dir=model_dir, device_str=device)
+
+    audio = pipe(
+        prompt=prompt,
+        seconds=seconds,
+        num_inference_steps=int(steps),
+        cfg_scale=float(cfg_scale),
+        sigma_shift=float(sigma_shift),
+        seed=int(seed),
     )
 
-    conversations, expected_tokens = build_conversation(
-        ambient_sound=ambient_sound,
-        duration_seconds=duration_seconds,
-        processor=processor,
-    )
-
-    batch = processor(conversations, mode="generation")
-    input_ids = batch["input_ids"].to(torch_device)
-    attention_mask = batch["attention_mask"].to(torch_device)
-
-    with torch.no_grad():
-        outputs = model.generate(
-            input_ids=input_ids,
-            attention_mask=attention_mask,
-            max_new_tokens=int(max_new_tokens),
-            audio_temperature=float(temperature),
-            audio_top_p=float(top_p),
-            audio_top_k=int(top_k),
-            audio_repetition_penalty=float(repetition_penalty),
-        )
-
-    messages = processor.decode(outputs)
-    if not messages or messages[0] is None:
-        raise RuntimeError("The model did not return a decodable audio result.")
-
-    audio = messages[0].audio_codes_list[0]
-    if isinstance(audio, torch.Tensor):
-        audio_np = audio.detach().float().cpu().numpy()
-    else:
-        audio_np = np.asarray(audio, dtype=np.float32)
-
-    if audio_np.ndim > 1:
-        audio_np = audio_np.reshape(-1)
+    audio_np = audio[0].detach().float().cpu().numpy()
+    if audio_np.ndim > 1 and audio_np.shape[0] == 1:
+        audio_np = audio_np.squeeze(0)
+    elif audio_np.ndim > 1:
+        audio_np = audio_np.T
     audio_np = audio_np.astype(np.float32, copy=False)
 
     elapsed = time.monotonic() - started_at
     status = (
         f"Done | elapsed: {elapsed:.2f}s | "
-        f"duration_seconds={float(duration_seconds):.0f}, expected_tokens={int(expected_tokens)}, "
-        f"max_new_tokens={int(max_new_tokens)}, "
-        f"audio_temperature={float(temperature):.2f}, audio_top_p={float(top_p):.2f}, "
-        f"audio_top_k={int(top_k)}, audio_repetition_penalty={float(repetition_penalty):.2f}"
+        f"duration={seconds:.1f}s, steps={int(steps)}, "
+        f"cfg_scale={float(cfg_scale):.2f}, sigma_shift={float(sigma_shift):.2f}, "
+        f"seed={int(seed)}"
     )
-    return (sample_rate, audio_np), status
+    return (pipe.sample_rate, audio_np), status
 
 
 def build_demo(args: argparse.Namespace):
@@ -182,23 +106,11 @@ def build_demo(args: argparse.Namespace):
       font-size: 22px;
       font-weight: 700;
       margin-bottom: 6px;
-      letter-spacing: 0.2px;
     }
     .app-subtitle {
       color: var(--muted);
       font-size: 14px;
       margin-bottom: 8px;
-    }
-    #output_audio {
-      padding-bottom: 12px;
-      margin-bottom: 8px;
-      overflow: hidden !important;
-    }
-    #output_audio > .wrap {
-      overflow: hidden !important;
-    }
-    #output_audio audio {
-      margin-bottom: 6px;
     }
     #run-btn {
       background: var(--accent);
@@ -206,139 +118,79 @@ def build_demo(args: argparse.Namespace):
     }
     """
 
-    with gr.Blocks(title="MOSS-SoundEffect Demo", css=custom_css) as demo:
+    with gr.Blocks(title="MOSS-SoundEffect v2.0", css=custom_css) as demo:
         gr.Markdown(
             """
             <div class="app-card">
-              <div class="app-title">MOSS-SoundEffect</div>
-              <div class="app-subtitle">Generate ambient sounds and sound effects from text descriptions.</div>
+              <div class="app-title">MOSS-SoundEffect v2.0</div>
+              <div class="app-subtitle">Text-to-audio diffusion (MossSoundEffectPipeline).</div>
             </div>
             """
         )
 
         with gr.Row(equal_height=False):
             with gr.Column(scale=3):
-                ambient_sound = gr.Textbox(
-                    label="Ambient Sound Description",
+                prompt = gr.Textbox(
+                    label="Prompt",
                     lines=8,
-                    placeholder="Example: Thunder rolls in the distance while heavy rain falls on a metal roof.",
+                    value="The crisp, rhythmic click-clack of fast typing on a mechanical keyboard.",
                 )
-                duration_seconds = gr.Slider(
-                    minimum=1,
-                    maximum=60,
-                    step=1,
-                    value=10,
+                seconds = gr.Slider(
+                    minimum=1, maximum=MAX_INFERENCE_SECONDS, step=0.1, value=10,
                     label="Duration (seconds)",
                 )
-
-                with gr.Accordion("Sampling Parameters (Audio)", open=True):
-                    temperature = gr.Slider(
-                        minimum=0.1,
-                        maximum=3.0,
-                        step=0.05,
-                        value=1.5,
-                        label="temperature",
-                    )
-                    top_p = gr.Slider(
-                        minimum=0.1,
-                        maximum=1.0,
-                        step=0.01,
-                        value=0.6,
-                        label="top_p",
-                    )
-                    top_k = gr.Slider(
-                        minimum=1,
-                        maximum=200,
-                        step=1,
-                        value=50,
-                        label="top_k",
-                    )
-                    repetition_penalty = gr.Slider(
-                        minimum=0.8,
-                        maximum=2.0,
-                        step=0.05,
-                        value=1.2,
-                        label="repetition_penalty",
-                    )
-                    max_new_tokens = gr.Slider(
-                        minimum=256,
-                        maximum=8192,
-                        step=128,
-                        value=DEFAULT_MAX_NEW_TOKENS,
-                        label="max_new_tokens",
-                    )
-
+                with gr.Accordion("Sampling Parameters", open=True):
+                    steps = gr.Slider(minimum=10, maximum=150, step=1, value=100, label="num_inference_steps")
+                    cfg_scale = gr.Slider(minimum=1.0, maximum=8.0, step=0.1, value=4.0, label="cfg_scale")
+                    sigma_shift = gr.Slider(minimum=0.0, maximum=10.0, step=0.1, value=5.0, label="sigma_shift")
+                    seed = gr.Number(value=0, label="seed", precision=0)
                 run_btn = gr.Button("Generate Sound Effect", variant="primary", elem_id="run-btn")
-
             with gr.Column(scale=2):
-                output_audio = gr.Audio(label="Output Audio", type="numpy", elem_id="output_audio")
+                output_audio = gr.Audio(label="Output Audio", type="numpy")
                 status = gr.Textbox(label="Status", lines=4, interactive=False)
 
         run_btn.click(
-            fn=lambda ambient_sound, duration_seconds, temperature, top_p, top_k, repetition_penalty, max_new_tokens: run_inference(
-                ambient_sound=ambient_sound,
-                duration_seconds=duration_seconds,
-                temperature=temperature,
-                top_p=top_p,
-                top_k=top_k,
-                repetition_penalty=repetition_penalty,
-                max_new_tokens=max_new_tokens,
-                model_path=args.model_path,
+            fn=lambda prompt, seconds, steps, cfg_scale, sigma_shift, seed: run_inference(
+                prompt=prompt,
+                seconds=seconds,
+                steps=steps,
+                cfg_scale=cfg_scale,
+                sigma_shift=sigma_shift,
+                seed=seed,
+                model_dir=args.model_dir,
                 device=args.device,
-                attn_implementation=args.attn_implementation,
             ),
-            inputs=[
-                ambient_sound,
-                duration_seconds,
-                temperature,
-                top_p,
-                top_k,
-                repetition_penalty,
-                max_new_tokens,
-            ],
+            inputs=[prompt, seconds, steps, cfg_scale, sigma_shift, seed],
             outputs=[output_audio, status],
         )
     return demo
 
 
 def main():
-    parser = argparse.ArgumentParser(description="MOSS-SoundEffect Gradio Demo")
-    parser.add_argument("--model_path", type=str, default=MODEL_PATH)
-    parser.add_argument("--device", type=str, default="cuda:0")
-    parser.add_argument("--attn_implementation", type=str, default=DEFAULT_ATTN_IMPLEMENTATION)
+    parser = argparse.ArgumentParser(description="MossSoundEffect Gradio Demo")
+    parser.add_argument("--model_dir", type=str, default=DEFAULT_MODEL_DIR)
+    parser.add_argument("--device", type=str, default=DEFAULT_DEVICE)
     parser.add_argument("--host", type=str, default="0.0.0.0")
     parser.add_argument("--port", type=int, default=7861)
+    parser.add_argument(
+        "--root_path",
+        type=str,
+        default=os.environ.get("GRADIO_ROOT_PATH"),
+        help="Mount path when serving behind a reverse proxy.",
+    )
     parser.add_argument("--share", action="store_true")
     args = parser.parse_args()
 
-    runtime_device = torch.device(args.device if torch.cuda.is_available() else "cpu")
-    runtime_dtype = torch.bfloat16 if runtime_device.type == "cuda" else torch.float32
-    args.attn_implementation = resolve_attn_implementation(
-        requested=args.attn_implementation,
-        device=runtime_device,
-        dtype=runtime_dtype,
-    ) or "none"
-    print(f"[INFO] Using attn_implementation={args.attn_implementation}", flush=True)
-
-    preload_started_at = time.monotonic()
-    print(
-        f"[Startup] Preloading backend: model={args.model_path}, device={args.device}, attn={args.attn_implementation}",
-        flush=True,
-    )
-    load_backend(
-        model_path=args.model_path,
-        device_str=args.device,
-        attn_implementation=args.attn_implementation,
-    )
-    print(
-        f"[Startup] Backend preload finished in {time.monotonic() - preload_started_at:.2f}s",
-        flush=True,
-    )
+    print(f"[Startup] Preloading backend: model_dir={args.model_dir}, device={args.device}", flush=True)
+    started = time.monotonic()
+    load_backend(model_dir=args.model_dir, device_str=args.device)
+    print(f"[Startup] Backend ready in {time.monotonic() - started:.2f}s", flush=True)
 
     demo = build_demo(args)
     demo.queue(max_size=16, default_concurrency_limit=1).launch(
         server_name=args.host,
         server_port=args.port,
+        root_path=args.root_path,
         share=args.share,
     )
