+++ moss_soundeffect_v2/finetuning/finetuning.py
@@ -16,7 +16,7 @@ import torch
 
 _HERE = Path(__file__).resolve().parent
 _PKG_DIR = _HERE.parent
-_PROJECT_DIR = _PKG_DIR.parent
+_PROJECT_DIR = _PKG_DIR.parent.parent
 sys.path.insert(0, str(_PROJECT_DIR))
 
 os.environ.setdefault("TOKENIZERS_PARALLELISM", "false")
